backend/app/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from cdp_langchain import CdpToolkit, CdpAgentkitWrapper
from cdp_agentkit_core.actions import CDP_ACTIONS
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI()

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Initialize CDP AgentKit
cdp = CdpAgentkitWrapper()
toolkit = CdpToolkit.from_cdp_agentkit_wrapper(cdp)

@app.get("/predictions")
async def get_predictions():
    # Get predictions using CDP AgentKit tools
    tools = toolkit.get_tools()
    # Implementation here
    pass

@app.post("/predictions/ai")
async def create_ai_prediction(asset: str):
    try:
        tools = toolkit.get_tools()
        # Use CDP AgentKit to create prediction
        # Implementation here
        pass
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))
